pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "utk-curio"
dynamic = ["version"]
description = "Curio: A dataflow-based framework for urban visual analytics"
authors = [{name = "UTK", email = "fabiom@uic.edu"}]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10,<3.12"

dependencies = [
  "requests==2.31.0",
  "pyarrow==15.0.0",
  "pillow==10.2.0",
  "rasterio==1.3.9",
  "python-dotenv==1.0.1",
  "Flask==3.1.0",
  "Flask-Caching==2.3.0",
  "fiona==1.9.6",
  "Flask-Migrate==4.0.5",
  "Flask-SQLAlchemy==3.1.1",
  "google-auth==2.27.0",
  "pandas==2.2.1",
  "altair>=5.0.0",
  "geopandas==1.0.1",
  "pytest==8.3.4",
  "pytest-playwright==0.7.0",
  "pytest-flask==1.3.0",
  "black==24.10.0"
]

[project.scripts]
curio = "utk_curio.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "utk_curio.__version__"}

[tool.setuptools.packages.find]
where = ["."]



